BIGAI ENGINE API SERVER TTS ALL LANGUAGES

+++ ENGINE/KEY_OPENAI.py
@@ -10,15 +10,21 @@ f = open(path+"account.txt", "r")
 key = f.read().strip()
 client = OpenAI(api_key=key)
 
-def generate_and_play(text, voice):
+def generate_and_play(text, voice, path=''):
     response = client.audio.speech.create(
         model="tts-1",
         voice=voice,
         input=text
     )
-    response.stream_to_file("oko.mp3")
-    playsound('oko.mp3')
-    os.remove('oko.mp3')
+    if path == '':
+        response.stream_to_file("oko.mp3")
+        playsound('oko.mp3')
+        os.remove('oko.mp3')
+    else:
+        response.stream_to_file(path)
+        playsound(path)
+
+
